# docs/api/swagger.yaml
basePath: /
definitions:
  github_com_zhian9_blogo-server_pkg_errors.Error:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      detail:
        description: 详细描述
        type: string
      id:
        description: 错误标识（前端友好）
        type: string
      status:
        description: 状态文本（如 "Not Found"）
        type: string
    type: object
...
  schema.Category:
    properties:
      article_count:
        description: 文章数量（非DB字段，查询时填充）
        type: integer
      created_at:
        description: 创建时间
        type: string
      id:
        description: 唯一ID
        type: string
      name:
        description: 分类名称（唯一）
        type: string
      sort:
        description: 排序值（越小越靠前）
        type: integer
      updated_at:
        description: 更新时间
        type: string
    type: object
  schema.CategoryForm:
    properties:
      name:
        description: 分类名称
        maxLength: 100
        type: string
      sort:
        description: 排序值
        minimum: 0
        type: integer
    required:
    - name
    type: object
...
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_zhian9_blogo-server_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8040
info:
  contact:
    email: lxyaa911@gmail.com
    name: 李星云 (lxy911)
    url: https://github.com/zhian9
  description: Modern SaaS Blog Platform — REST API for content management, user system,
    RBAC, analytics, and more.
  license:
    name: MIT
    url: https://github.com/zhian9/Blogo/blob/master/LICENSE
  termsOfService: https://github.com/zhian9/Blogo
  title: Blogo API
  version: v1.0.0
paths:
...
  /api/v1/categories:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        type: integer
      - description: Category name (fuzzy)
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Category'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Query category list
      tags:
      - CategoryAPI
    post:
      parameters:
      - description: Category form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.CategoryForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - CategoryAPI
  /api/v1/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - CategoryAPI
    get:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Category'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Get category by ID
      tags:
      - CategoryAPI
    put:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.CategoryForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - CategoryAPI
  /api/v1/categories/all:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Category'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Get all categories (public)
      tags:
      - CategoryAPI
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your JWT token>". Login via /api/v1/login to obtain
      a token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"